refactor: withQuery 따로 공통 함수로 뺌
